Refactored CLI menu functionality
